Apply Boxer compilation fixes

+++ src/hardware/serialport/misc_util.cpp
@@ -227,7 +227,7 @@ bool TCPClientSocket::Putchar(Bit8u data) {
 }
 
 bool TCPClientSocket::SendArray(Bit8u* data, Bitu bufsize) {
-	if(SDLNet_TCP_Send(mysock, data, bufsize)!=bufsize) {
+	if(SDLNet_TCP_Send(mysock, data, bufsize)!=(int)bufsize) {
 		isopen=false;
 		return false;
 	}
@@ -241,7 +241,7 @@ bool TCPClientSocket::SendByteBuffered(Bit8u data) {
 		sendbuffer[sendbufferindex]=data;
 		sendbufferindex=0;
 		
-		if(SDLNet_TCP_Send(mysock, sendbuffer, sendbuffersize)!=sendbuffersize) {
+		if(SDLNet_TCP_Send(mysock, sendbuffer, sendbuffersize)!=(int)sendbuffersize) {
 			isopen=false;
 			return false;
 		}
@@ -273,7 +273,7 @@ bool TCPClientSocket::SendArrayBuffered(Bit8u* data, Bitu bufsize) {
 void TCPClientSocket::FlushBuffer() {
 	if(sendbufferindex) {
 		if(SDLNet_TCP_Send(mysock, sendbuffer,
-			sendbufferindex)!=sendbufferindex) {
+			sendbufferindex)!=(int)sendbufferindex) {
 			isopen=false;
 			return;
 		}
